Moved Windows-only Permission function to Sonarr.Windows
