+++ src/NzbDrone.Common/Disk/IDiskProvider.cs
@@ -10,6 +10,7 @@ namespace NzbDrone.Common.Disk
     {
         long? GetAvailableSpace(string path);
         void InheritFolderPermissions(string filename);
+        void SetEveryonePermissions(string filename);
         void SetPermissions(string path, string mask);
         void CopyPermissions(string sourcePath, string targetPath);
         long? GetTotalSize(string path);
@@ -40,7 +41,6 @@ namespace NzbDrone.Common.Disk
         bool IsFileLocked(string path);
         string GetPathRoot(string path);
         string GetParentFolder(string path);
-        void SetPermissions(string filename, WellKnownSidType accountSid, FileSystemRights rights, AccessControlType controlType);
         FileAttributes GetFileAttributes(string path);
         void EmptyFolder(string path);
         string GetVolumeLabel(string path);
