+++ src/NzbDrone.Mono/Disk/DiskProvider.cs
@@ -56,20 +56,12 @@ namespace NzbDrone.Mono.Disk
 
         public override void InheritFolderPermissions(string filename)
         {
-            Ensure.That(filename, () => filename).IsValidPath();
 
-            try
-            {
-                var fs = File.GetAccessControl(filename);
-                fs.SetAccessRuleProtection(false, false);
-                File.SetAccessControl(filename, fs);
-            }
-            catch (NotImplementedException)
-            {
-            }
-            catch (PlatformNotSupportedException)
-            {
-            }
+        }
+
+        public override void SetEveryonePermissions(string filename)
+        {
+            
         }
 
         public override void SetPermissions(string path, string mask)
